Fix rendering at low zoom levels
At low zoom levels, the longitude coordinates of map bounds returned
by getBounds() may be outside of the normal ±180° range due to horizontal
wrapping. In order for the location filter to render correctly, it is
necessary to preserve these longitudes by passing true as the third
argument to all LatLng constructors.